DumpGenerator: Move "debug" outputs to stdout for "clean" stderr
I redirect StdOut to /dev/null in scheduled wikidumps, to only receive a message when there is an actual error. These two debug outputs are still written to StdErr, which leads to me getting emails with just a single "." in them. I've moved them to regular stdout, but alternatively they could also be removed or replaced with a proper log message?