Enhance the stability of the image dump and make it resumable (#88)
---

- Introduce `sha1File`
- save more metadata(`size`, `sha1`) into `images.txt`
- feat: better file dump:
  - validate image's size and sha1
  - show progress
  - better resume
    > Improved the resume mechanism. (fix: #15)
First check whether the `file` and `file.desc` exist, and then check
whether the `size` and `sha1` of the file correspond to `images.txt`. If
any check fails, the file and the .desc of the file will be downloaded
again. If all pass, the download of this file is skipped.
You can even delete random pictures and .desc files and try to resume
again.
  - pre-work for incremental image dump
  - remove `start` param from `generatorImageDump()`
> the images resume mechanism has changed. we don't need `start` for
resuming anymore.
- other minor improvements